Create ICT Project
Aerodome works like an actual air defense system. it has an interceptor to intercept missiles , it has a 650 km range, 700 km detection range, mach 8 speed and reaction time of 15s. it as to read an excel file to train itself for the parameters that user will enter  after the model learns from the excel file